refactor(logstream): reduce visibility of sequencer internals
